LayoutTests/fast/viewport/viewport-131-expected.txt

ALERT: viewport size 320x352 scale 1 with limits [1, 1] and userScalable false
